Struct an_rope::RopeSlice

pub struct RopeSlice<'a> { /* fields omitted */ }

An immutable borrowed slice of a Rope.

A RopeSlice represents an immutable borrowed slice of some or all the characters in a Rope.

Methods

impl<'a> RopeSlice<'a>

pub fn chars(&'a self) -> Box<Iterator<Item = char> + 'a>

pub fn char_indices(&'a self) -> Box<Iterator<Item = (usize, char)> + 'a>

pub fn bytes(&'a self) -> Box<Iterator<Item = u8> + 'a>

pub fn split_whitespace(&'a self) -> Box<Iterator<Item = &'a str> + 'a>

pub fn new(node: &'a Node, range: Range<usize>) -> Self

pub fn len(&self) -> usize

pub fn is_empty(&self) -> bool

Returns true if this RopeSlice is empty.

Examples

A RopeSlice into an empty rope should be empty:

use an_rope::Rope;
let an_empty_rope = Rope::new();
assert!(an_empty_rope.slice(0..0).is_empty());

A RopeSlice that contains no characters should be empty, even fi the sliced Rope is not empty:

use an_rope::Rope;
let an_rope = Rope::from("a string that is not empty");
assert!(an_rope.slice(0..0).is_empty());

A RopeSlice with characters should not be empty:

use an_rope::Rope;
let an_rope = Rope::from("a string");
assert!(!an_rope.slice(0..5).is_empty());